requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0

# Backend API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Supabase
supabase>=2.0.0
postgrest>=0.10.0

# Supabase
supabase>=2.0.0
postgrest>=0.10.0

# Supabase
supabase>=2.0.0
postgrest>=0.10.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
spacy>=3.6.0

# Explainability
shap>=0.42.0
lime>=0.2.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# News Scraping
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
requests>=2.31.0

# MLOps
wandb>=0.15.0
mlflow>=2.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
pre-commit>=3.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Streamlit (for prototyping)
streamlit>=1.25.0
